Abstracted grapher file into smaller modules
