+++ gui/sidebar.js
@@ -1,56 +1,13 @@
 import Expression from "../calc/expression.js";
-import { functions, variables, addFunction, render, selectedFunction, setSelectedFunction, clearCache } from "./grapher.js";
+import { functions, variables, addFunction } from "./function.js";
+import { clearCache } from "./cache.js";
+import { functionColors } from "./util.js";
+import { render } from "./grapher.js";
 
-const functionColors = {
-    red: {
-        function: '#E57373',
-        focus: '#FFEBEE',
-    },
-    magenta: {
-        function: '#F06292',
-        focus: '#FCE4EC',
-        
-    },
-    purple: {
-        function: '#BA68C8',
-        focus: '#F3E5F5',
-    },
-    darkblue: {
-        function: '#7986CB',
-        focus: '#E8EAF6',
-    },
-    lightblue: {
-        function: '#4FC3F7',
-        focus: '#E1F5FE',
-    },
-    cyan: {
-        function: '#4DD0E1',
-        focus: '#E0F7FA',
-    },
-    teal: {
-        function: '#009688',
-        focus: '#E0F2F1',
-    },
-    green: {
-        function: '#81C784',
-        focus: '#E8F5E9',
-    },
-    yellow: {
-        function: '#FFF176',
-        focus: '#FFFDE7',
-    },
-    orange: {
-        function: '#FFB74D',
-        focus: '#FFF8E1',
-    },
-    amber: {
-        function: '#FF8A65',
-        focus: '#FBE9E7',
-    },
-    bluegrey: {
-        function: '#90A4AE',
-        focus: '#ECEFF1',
-    }
+export const selectedFunction = -1;
+
+function setSelectedFunction(index) {
+    selectedFunction = index;
 }
 
 const addInputButton = document.getElementById('add-input-button');
@@ -93,12 +50,8 @@ function updateInputs() {
             } catch (err) {
                 symbols[i].src = selected ? '../assets/caution_selected.png' : '../assets/caution.png';
             }
-        } 
+        }
 
-
-            
-        
-        
         functions[i].expression = value;
     }
     
@@ -106,7 +59,7 @@ function updateInputs() {
 }
 
 
-export function addInput() {
+function addInput() {
     let index = functions.length;
     let values = Object.values(functionColors);
     let rand = Math.floor(Math.random() * values.length);
